refactor: refactor ID generator structure for better separation of concerns
