+++ src/main/kotlin/com/dobby/backend/application/usecase/experiment/CreateExperimentPostUseCase.kt
@@ -2,7 +2,7 @@ package com.dobby.backend.application.usecase.experiment
 
 import com.dobby.backend.application.usecase.UseCase
 import com.dobby.backend.domain.exception.*
-import com.dobby.backend.domain.gateway.IdGeneratorGateway
+import com.dobby.backend.domain.IdGenerator
 import com.dobby.backend.domain.gateway.experiment.ExperimentPostGateway
 import com.dobby.backend.domain.gateway.member.MemberGateway
 import com.dobby.backend.domain.model.experiment.ApplyMethod
@@ -21,7 +21,7 @@ import java.time.LocalDate
 class CreateExperimentPostUseCase(
     private val experimentPostGateway: ExperimentPostGateway,
     private val memberGateway: MemberGateway,
-    private val idGeneratorGateway: IdGeneratorGateway
+    private val idGenerator: IdGenerator
 ) : UseCase<CreateExperimentPostUseCase.Input, CreateExperimentPostUseCase.Output> {
 
     data class Input(
@@ -88,7 +88,7 @@ class CreateExperimentPostUseCase(
         validateMemberRole(member)
 
         val targetGroup = TargetGroup.newTargetGroup(
-            id = idGeneratorGateway.generateId(),
+            id = idGenerator.generateId(),
             startAge = input.targetGroupInfo.startAge,
             endAge = input.targetGroupInfo.endAge,
             genderType = input.targetGroupInfo.genderType,
@@ -96,14 +96,14 @@ class CreateExperimentPostUseCase(
         )
 
         val applyMethod = ApplyMethod.newApplyMethod(
-            id = idGeneratorGateway.generateId(),
+            id = idGenerator.generateId(),
             phoneNum = input.applyMethodInfo.phoneNum,
             formUrl = input.applyMethodInfo.formUrl,
             content = input.applyMethodInfo.content
         )
 
         val experimentPost = ExperimentPost.newExperimentPost(
-            id = idGeneratorGateway.generateId(),
+            id = idGenerator.generateId(),
             member = member,
             targetGroup = targetGroup,
             applyMethod = applyMethod,
@@ -126,7 +126,7 @@ class CreateExperimentPostUseCase(
 
         val experimentImages = input.imageListInfo.images.map { imageUrl ->
             ExperimentImage.newExperimentImage(
-                id = idGeneratorGateway.generateId(),
+                id = idGenerator.generateId(),
                 experimentPost = null,
                 imageUrl = imageUrl
             )
