+++ src/main/kotlin/com/dobby/backend/domain/IdGenerator.kt
@@ -1,0 +1,5 @@
+package com.dobby.backend.domain
+
+interface IdGenerator {
+    fun generateId(): String
+}
