--- src/main/kotlin/com/dobby/backend/domain/gateway/IdGeneratorGateway.kt
@@ -1,5 +1,0 @@
-package com.dobby.backend.domain.gateway
-
-interface IdGeneratorGateway {
-    fun generateId(): String
-}
