+++ src/test/kotlin/com/dobby/backend/application/usecase/member/CreateResearcherUseCaseTest.kt
@@ -1,6 +1,6 @@
 package com.dobby.backend.application.usecase.member
 
-import com.dobby.backend.domain.gateway.IdGeneratorGateway
+import com.dobby.backend.domain.IdGenerator
 import com.dobby.backend.domain.gateway.auth.TokenGateway
 import com.dobby.backend.domain.gateway.member.MemberGateway
 import com.dobby.backend.domain.gateway.member.ResearcherGateway
@@ -19,9 +19,9 @@ class CreateResearcherUseCaseTest : BehaviorSpec({
     val memberGateway: MemberGateway = mockk()
     val researcherGateway: ResearcherGateway = mockk()
     val tokenGateway: TokenGateway = mockk()
-    val idGeneratorGateway: IdGeneratorGateway = mockk()
+    val idGenerator: IdGenerator = mockk()
 
-    val createResearcherUseCase = CreateResearcherUseCase(memberGateway, researcherGateway, tokenGateway, idGeneratorGateway)
+    val createResearcherUseCase = CreateResearcherUseCase(memberGateway, researcherGateway, tokenGateway, idGenerator)
 
     given("유효한 입력을 받았을 때") {
         val input = CreateResearcherUseCase.Input(
@@ -59,7 +59,7 @@ class CreateResearcherUseCaseTest : BehaviorSpec({
         val accessToken = "mock-access-token"
         val refreshToken = "mock-refresh-token"
 
-        every { idGeneratorGateway.generateId() } returns "1"
+        every { idGenerator.generateId() } returns "1"
         every { researcherGateway.save(any()) } returns savedResearcher
         every { memberGateway.save(any()) } returns savedMember
         every { tokenGateway.generateAccessToken(savedMember) } returns accessToken
