+++ src/test/kotlin/com/dobby/backend/infrastructure/s3/S3PreSignedUrlProviderTest.kt
@@ -3,7 +3,7 @@ package com.dobby.backend.infrastructure.s3
 import com.amazonaws.services.s3.AmazonS3
 import com.amazonaws.services.s3.model.GeneratePresignedUrlRequest
 import com.dobby.backend.domain.exception.InvalidRequestValueException
-import com.dobby.backend.domain.gateway.IdGeneratorGateway
+import com.dobby.backend.domain.IdGenerator
 import com.dobby.backend.infrastructure.config.properties.S3Properties
 import io.kotest.core.spec.style.BehaviorSpec
 import org.mockito.Mockito.*
@@ -19,12 +19,12 @@ class S3PreSignedUrlProviderTest : BehaviorSpec({
     val amazonS3Client = mock(AmazonS3::class.java)
     val s3Properties = mock(S3Properties::class.java)
     val s3Config = mock(S3Properties.S3::class.java)
-    val idGeneratorGateway = mock(IdGeneratorGateway::class.java)
+    val idGenerator = mock(IdGenerator::class.java)
 
     `when`(s3Properties.s3).thenReturn(s3Config)
     `when`(s3Config.bucket).thenReturn("test-bucket")
     `when`(s3Properties.s3.bucket).thenReturn("test-bucket")
-    val provider = S3PreSignedUrlProvider(amazonS3Client, idGeneratorGateway, s3Properties)
+    val provider = S3PreSignedUrlProvider(amazonS3Client, idGenerator, s3Properties)
 
     given("이미지 파일 이름이 주어지고") {
         val imageName = "test_image.jpg"
@@ -52,8 +52,8 @@ class S3PreSignedUrlProviderTest : BehaviorSpec({
     }
 
     given("generateId 함수가 호출되면") {
-        `when`(idGeneratorGateway.generateId()).thenReturn("test-tsid")
-        val tsid = idGeneratorGateway.generateId()
+        `when`(idGenerator.generateId()).thenReturn("test-tsid")
+        val tsid = idGenerator.generateId()
 
         then("TSID가 생성된다") {
             assertNotNull(tsid)
